ws-client: Improve subscriptions management
subscribe() method gets a new `subscriber` parameter to identify
a subscription and makes it possible to unsubscribe it later by
a call to unsubscribe() with the same parameters.

Only one path:method subscription is possible for each subscriber.
Actual subscribe/unsubscribe RPC calls to the broker are made only
upon first/last subscribe/unsubscribe of a particular path:method.